+++ clang-tools-extra/test/clang-tidy/checkers/modernize/use-ranges.cpp
@@ -43,6 +43,24 @@ void Positives() {
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
   // CHECK-FIXES: std::ranges::reverse(I);
 
+  auto LogicalEnd = std::unique(I.begin(), I.end());
+  // CHECK-MESSAGES: :[[@LINE-1]]:21: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: auto LogicalEnd = std::ranges::unique(I).begin();
+
+  bool AlreadyUnique = std::unique(I.begin(), I.end()) == I.end();
+  // CHECK-MESSAGES: :[[@LINE-1]]:24: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: bool AlreadyUnique = std::ranges::unique(I).begin() == I.end();
+
+  auto LogicalEndWithPred =
+      std::unique(I.begin(), I.end(), [](int A, int B) { return A == B; });
+  // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: auto LogicalEndWithPred =
+  // CHECK-FIXES-NEXT: std::ranges::unique(I, [](int A, int B) { return A == B; }).begin();
+
+  std::unique(I.begin(), I.end());
+  // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: std::ranges::unique(I);
+
   std::includes(I.begin(), I.end(), I.begin(), I.end());
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
   // CHECK-FIXES: std::ranges::includes(I, I);
@@ -79,9 +97,18 @@ void Positives() {
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
   // CHECK-FIXES: std::ranges::find(I, 5);
 
+  using std::unique;
+  auto LogicalEndFromUsing = unique(I.begin(), I.end());
+  // CHECK-MESSAGES: :[[@LINE-1]]:30: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: auto LogicalEndFromUsing = std::ranges::unique(I).begin();
+
   my_std::find(I.begin(), I.end(), 6);
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
   // CHECK-FIXES: std::ranges::find(I, 6);
+
+  auto LogicalEndFromNamespaceAlias = my_std::unique(I.begin(), I.end());
+  // CHECK-MESSAGES: :[[@LINE-1]]:39: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: auto LogicalEndFromNamespaceAlias = std::ranges::unique(I).begin();
 }
 
 void Reverse(){
@@ -98,6 +125,10 @@ void Reverse(){
   std::find(I.rbegin(), I.rend(), 0);
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm
   // CHECK-FIXES: std::ranges::find(std::views::reverse(I), 0);
+
+  auto ReverseLogicalEnd = std::unique(I.rbegin(), I.rend());
+  // CHECK-MESSAGES: :[[@LINE-1]]:28: warning: use a ranges version of this algorithm
+  // CHECK-FIXES: auto ReverseLogicalEnd = std::ranges::unique(std::views::reverse(I)).begin();
 
   std::equal(std::rbegin(I), std::rend(I), J.begin(), J.end());
   // CHECK-MESSAGES: :[[@LINE-1]]:3: warning: use a ranges version of this algorithm